delete confirm added

+++ src/main/webapp/WEB-INF/views/category/all_category.jsp
@@ -25,7 +25,7 @@
     <tr>
         <td>${category.name}</td>
         <td><a href="/api/category/edit/${category.id}">Edytuj</a> </td>
-        <td><a href="/api/category/delete/${category.id}">Usuń</a> </td>
+        <td><a class="confirm" href="/api/category/delete/${category.id}">Usuń</a> </td>
     </tr>
     </c:forEach>
 </table>
